% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{What time the people arrive}

\item{ServiceTimes}{The time they took at the counter with their service}

\item{NumServers}{How many servers there are available}
}
\value{
data.frame of 4 variables about the Arrivals, ServiceBegins, ChosenServer, Service Ends
}
\description{
Multiserver
}
\examples{
Multiserver(bank$arrival_time,bank$service_time)

}
