% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Data}
\format{
\subsection{A data frame with 100 rows and 2 variables}{

\describe{
\item{arrival_time}{The time of the individuals arrival}
\item{service_time}{The time the time the individual is served}
...
}
}
}
\source{
STAT1378 Assignment
data(bank)
}
\usage{
bank
}
\description{
Data implementing a first come, first serve queuing system in a bank.
}
\keyword{datasets}
