% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 100 rows and 3 columns.
}
\source{
\url{https://www.npmjs.com/package/multiserver}
}
\usage{
bank
}
\description{
simulates the comings and goings of customers through the system, along with the time when each customer will arrive and their service times.
}
\details{
Report ...

A data frame with 100 rows and 2 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
\keyword{datasets}
