% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{bank}
\format{
\subsection{'bank'}{

A data frame with 100 rows and 3 variables:
\describe{
\item{arrival_time}{arrival time}
\item{service_time}{service time}
...
}
}
}
\usage{
bank
}
\description{
Dataset for customer queuing
}
\keyword{datasets}
