% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Arrival time and service duration of customers}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{arrival_time}{Arrival time of customer}
\item{service_time}{duration of service}
...
}
}
\usage{
bank
}
\description{
A dataset containing the arrival time and service edit
}
\keyword{datasets}
