% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank_data.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Data Set}
\format{
An object of class \code{data.frame} with 100 rows and 2 columns.
}
\usage{
bank
}
\description{
A data set containing the arrival times and service times of customers within a bank.
}
\keyword{datasets}
