% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Data about customers arrival time and service duration}
\format{
A tibble with 100 rows and 3 variables:
\describe{
\item{}{dbl index}
\item{arrival_time}{dbl customer's arrival time}
\item{service_time}{dbl customer's service duration}

}
}
\source{
{given by macquarie university}
}
\usage{
bank
}
\description{
Customers arrival time and their service duration is noted
}
\examples{
data(bank)
}
\keyword{datasets}
