% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Data}
\format{
\code{bank}
A data frame with 100 rows and 2 columns:
\describe{
\item{arrival}{time}
\item{service}{time}
}
}
\usage{
bank
}
\description{
Data taken from the arrival times and service times of customers at a bank
}
\keyword{datasets}
