% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{Data to showcase our service time structures}
\format{
A data frame with 100 observations and 3 variables:
\describe{
\item{customer}{the customer number}
\item{arrival_time}{the time the customer arrives}
\item{service_time}{the time it takes to serve the customer}
}

@source {Provided data set to serve as an example of simulating service in a store.}
}
\usage{
bank
}
\description{
Contains the customer number and their specific arrival and service time.
}
\examples{
data(bank)
}
\keyword{datasets}
