% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Data For Customers}
\format{
A data frame with 200 rows and 2 variables:
\describe{
\item{arrival_time}{arrival time, in minutes}
\item{service_time}{service time, in minutes}
...
}
}
\usage{
bank
}
\description{
A dataset containing the arrival times and service times of a 100 customers who visit the bank.
}
\keyword{datasets}
