% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiserver.R
\name{multiserver}
\alias{multiserver}
\title{Multiserver}
\usage{
multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{The time when the individual arrives to be served}

\item{ServiceTimes}{The time it takes to be serves}

\item{NumServers}{How many servers are there}
}
\value{
A tibble with their time of arrival, the time they are served, who they are served by and when their service ends
}
\description{
The function simulates a first come, first serve queuing system when fate
is the arrival
}
\examples{
multiserver(Arrivals = 1, ServiceTimes = 2, NumServers = 1)
}
