% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiserver.R
\name{multiserver}
\alias{multiserver}
\title{Multiserver}
\usage{
multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{The time when the individual arrives to be served}

\item{ServiceTimes}{The time it takes to be served}

\item{NumServers}{The number of servers there are}
}
\value{
A tibble with their time of arrival, the time they are served, who served them and when their service ends
}
\description{
The function simulates a first come, first serve queuing system when arrival and service times are predetermined.
}
\examples{
multiserver(Arrivals = 1, ServiceTimes = 2, NumServers = 1)
}
