% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{multiserver}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{indicates the arrival times of customers}

\item{ServiceTimes}{indicates the time it takes to service a customer}

\item{NumServers}{indicates how many servers are available to the customers}
}
\value{
a tibble which displays arrival and start times, along with the end times at the chosen server
}
\description{
simulates how customers will go through a first come, first serve queuing system as long as “fate” has already decided when each customer will arrive and their service times.
}
\examples{
Multiserver(bank$arrival_time, bank$service_time, NumServers = 2)

}
