% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{Tells us the time that the customer arrives}

\item{ServiceTimes}{Tells us the time taken for the service}

\item{NumServers}{Tells us the number of servers that are available}
}
\value{
Tibble of Arrivals, ServiceBegins, ChosenServer, ServiceEnds
}
\description{
Multiserver
}
\examples{
Multiserver(Arrivals = 1, ServiceTimes = 2)
}
