% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver 2}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{Shows the arrival time and each customer}

\item{ServiceTimes}{Displays the service times of each customer}

\item{NumServers}{Dispays the number of people to be serviced}
}
\value{
Displays the service time intervals
}
\description{
Multiserver 2
}
