% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Bank Queuing system}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{includes the number of customers, the time they arrive, and their service duration}

\item{ServiceTimes}{Includes the start time of service and the time where the service ends}

\item{NumServers}{the number of servers in the bank simulation}
}
\value{
a tibble with columns: Arrivals, ServiceBegins, ChosenServer, ServiceEnds
}
\description{
Bank Queuing system
}
\examples{
Multiserver(23, 3, 1)
}
