% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver.}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{vector; In seconds.}

\item{ServiceTimes}{vector; In seconds.}

\item{NumServers}{vector; In seconds.}
}
\value{
A data frame
}
\description{
Multiserver.
}
\examples{
Multiserver(6,5,1)
}
