% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{denotes the time at which the customer arrives}

\item{ServiceTimes}{denotes the customer's service duration}

\item{NumServers}{denotes the total number of servers available for the shift}
}
\value{
a tibble that displays the customer's arrival time, the chosen server, service starting and ending time
}
\description{
Feed customers through a multiserver queuing system to determine each customer's service transition times.
}
\examples{
Multiserver(bank$arrival_time, bank$service_time, 3)
Multiserver(bank$arrival_time, bank$service_time, 2)
Multiserver(bank$arrival_time, bank$service_time, 1)

}
