% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver Package Title}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{A numeric vector of customer arrival times.}

\item{ServiceTimes}{A numeric vector of customer service times.}

\item{NumServers}{An integer specifying the number of servers in the system (default is 1).}
}
\value{
\item{Arrivals}{The arrival times of the customers.}
\item{ServiceBegins}{The times when each customer's service begins.}
\item{ChosenServer}{The server assigned to each customer.}
\item{ServiceEnds}{The times when each customer's service ends.}
}
\description{
the package provides tools for simulating "first come first serve" queing system with multiple servers, meaning it simulates multiserver queing systems based on customer arrival and service time data, which will calculate individual customer's waiting times, service times, and service end times.
}
\details{
Multiserver Package

A package for simulating multi-server queue systems.
}
\examples{
# Example usage:
arrivals <- c(1, 2, 4, 6)
service_times <- c(3, 2, 5, 1)
NumServers <- 2
Multiserver(arrivals, service_times, NumServers)

}
