% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Simulate service times}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{The arrival time of the customer wanting to be served}

\item{ServiceTimes}{The amount of time a customer is served for}

\item{NumServers}{The number of servers available to work in the service system}
}
\value{
The server chosen to serve a particular customer along with the beginning and end of service with the chosen server.
}
\description{
This function is useful in ensuring customers transitions through a queuing system given their arrival and service times.
}
\examples{
Multiserver(120, 30, 3)
}
