% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Calculating Service Times For Customers At A Bank}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{text string; Time the customer first arrives.}

\item{ServiceTimes}{text string; The total duration of service, i.e., difference between the start time of service and the end time of service.}

\item{NumServers}{text string; The number of servers available to serve customers}
}
\value{
data frame
}
\description{
This package will outline which server is available first for a certain customer.
}
